BROKEN transport commit with 2 new URL handlers
This is a broken commit because I need to work from a different machine afterwards. It adds two new URLs and their handler functions to ui/urls.py and ui/views.py, adapted from Indivo. The handler have yet to be adapted. Also new is a new template for record selection, which will have parts of it spun out into a base template.