Spring cleaning indivo_client_py
Removed the config files as we are not going to re-create the api (I guess). Removed the create_api.py script. And removed most of the generated api calls, leaving only those that could potentially be used. Maybe they should all go and we should just use a generic api call where we manually specify the URL?
- remove carenet stuff
- remove autoshare stuff
- remove document POST/PUT
- remove document relations
- remove app document handling stuff
- remove app handling stuff
- remove Indivo reporting stuff
- remove document deletion calls
- remove audit calls
- remove notification calls
- remove a lot of individual calls
- remove coding system lookup